+++ app/src/main/java/com/example/weitblickapp_android/ui/news/NewsListFragment.java
@@ -315,8 +315,8 @@ public class NewsListFragment extends ListFragment implements AbsListView.OnScro
 
                             locationObject = responseObject.getJSONObject("location");
 
-                            float lat = locationObject.getLong("lat");
-                            float lng = locationObject.getLong("lng");
+                            double lat = locationObject.getLong("lat");
+                            double lng = locationObject.getLong("lng");
                             String name = locationObject.getString("name");
                             String address = locationObject.getString("address");
                             String descriptionLocation = locationObject.getString("description");
